pyproject.toml
[tool.poetry]
name = "eventide-rag"
version = "0.1.0"
description = ""
authors = ["Justin <litch@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
langchain = "^0.1.17"
langchain-openai = "^0.1.4"
pinecone-client = "^3.2.2"
beautifulsoup4 = "^4.12.3"
langchain-community = "^0.0.36"
unstructured = {extras = ["md"], version = "^0.13.6"}
python-dotenv = "^1.0.1"
streamlit = "^1.34.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
